[ZEPPELIN-6086] Remove Spark Shims and unofficial support for Spark 4.0
### What is this PR for?

Add support for Spark 4 by removing the spark shims. There is only one version of shims now, and they are compatible with the current Spark 4.0 preview releases, so we can just remove the shim concept.

### What type of PR is it?
Improvement

### Todos
* [x] - Disable 4.0 tests in CI and re-enable supported version check

### What is the Jira issue?

ZEPPELIN-6086

### How should this be tested?
A CI stage is added for Spark 4.0 tests, but these will need to be disabled until 4.0 is officially supported.

### Screenshots (if appropriate)

### Questions:
* Does the license files need to update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? Maybe


Closes #4825 from Kimahriman/spark-4.0.
