+++ .gitignore
@@ -142,3 +142,6 @@ tramp
 
 # pyenv file
 .python-version
+
+# python venv
+.venv
