+++ spark/pom.xml
@@ -55,8 +55,7 @@
         <module>spark-scala-parent</module>
         <module>scala-2.12</module>
         <module>scala-2.13</module>
-        <module>spark-shims</module>
-        <module>spark3-shims</module>
+        <module>spark-common</module>
     </modules>
 
     <build>
