+++ spark/scala-2.13/src/main/scala/org/apache/zeppelin/spark/SparkZeppelinContext.scala
@@ -33,7 +33,7 @@ import scala.collection.JavaConverters._
   * ZeppelinContext for Spark
   */
 class SparkZeppelinContext(val sc: SparkContext,
-                           val sparkShims: SparkShims,
+                           val sparkUtils: SparkUtils,
                            val hooks2: InterpreterHookRegistry,
                            val maxResult2: Int) extends ZeppelinContext(hooks2, maxResult2) {
 
@@ -65,7 +65,7 @@ class SparkZeppelinContext(val sc: SparkContext,
 
   override def getInterpreterClassMap: util.Map[String, String] = interpreterClassMap.asJava
 
-  override def showData(obj: Any, maxResult: Int): String = sparkShims.showDataFrame(obj, maxResult, interpreterContext)
+  override def showData(obj: Any, maxResult: Int): String = sparkUtils.showDataFrame(obj, maxResult, interpreterContext)
 
   /**
    * create paragraph level of dynamic form of Select with no item selected.
@@ -237,6 +237,6 @@ class SparkZeppelinContext(val sc: SparkContext,
   }
 
   def getAsDataFrame(name: String): Object = {
-    sparkShims.getAsDataFrame(get(name).toString)
+    sparkUtils.getAsDataFrame(get(name).toString)
   }
 }
